missed some doubles

+++ src/hrtf/check.c
@@ -3,7 +3,7 @@
 #include "mysofa.h"
 #include "tools.h"
 
-static int compareValues(struct MYSOFA_ARRAY *array, double *compare,
+static int compareValues(struct MYSOFA_ARRAY *array, float *compare,
 		int elements) {
 	int i;
 	if (array->values == NULL || array->elements != elements)
@@ -14,10 +14,9 @@ static int compareValues(struct MYSOFA_ARRAY *array, double *compare,
 	return 1;
 }
 
-static double array000[] = { 0, 0, 0 };
-static double array001[] = { 0, 0, 1 };
-// static double array0901[] = { 0, 90, 1 };
-static double array100[] = { 1, 0, 0 };
+static float array000[] = { 0, 0, 0 };
+static float array001[] = { 0, 0, 1 };
+static float array100[] = { 1, 0, 0 };
 
 int mysofa_check(struct MYSOFA_HRTF *hrtf) {
